app/requirements.txt
fastapi
uvicorn
motor                # Async MongoDB driver
httpx
nltk
scikit-learn
pdfminer.six
python-docx
newspaper3k
pydantic
